refactor: extract response cycle logic
