+++ src/agent/core/index.ts
@@ -5,3 +5,4 @@ export * from './ToolConfig';
 export * from './ToolState';
 export * from './ResponseUsage';
 export * from './IAgent';
+export * from './ResponseCycle';
